athena_api/requirements-api.txt
fastapi
uvicorn
transformers
langchain
faiss-cpu
sentence-transformers
torch
scikit-learn
google-cloud-storage
joblib
gcsfs
pydantic